organized map structure and added videostate
