+++ ViAn/GUI/Analysis/analysisslider.cpp
