+++ ViAn/GUI/Analysis/analysisslider.h
