+++ ViAn/GUI/Project/recentprojectdialog.h
@@ -11,7 +11,7 @@
 #include <QLabel>
 
 #include "Project/recentproject.h"
-#include "GUI/projectdialog.h"
+#include "GUI/Project/projectdialog.h"
 
 class RecentProjectDialog : public QDialog {
     Q_OBJECT
