+++ ViAn/GUI/TreeItems/tagitem.h
@@ -1,0 +1,16 @@
+#ifndef TAGITEM_H
+#define TAGITEM_H
+#include "treeitem.h"
+#include "Project/Analysis/tag.h"
+#include "itemstates.h"
+class TagItem : public TreeItem {
+    Tag* m_tag;
+    VideoState state;
+public:
+    TagItem(Tag *tag);
+    Tag *get_tag();
+    void remove();
+    void rename();
+};
+
+#endif // TAGITEM_H
