+++ ViAn/GUI/TreeItems/treeitem.cpp
@@ -1,0 +1,6 @@
+#include "treeitem.h"
+
+TreeItem::TreeItem(int type) : QTreeWidgetItem(type) {
+    setFlags(Qt::ItemIsEditable | Qt::ItemIsEnabled | Qt::ItemIsSelectable);
+}
+
