+++ ViAn/GUI/Video/fpsdialog.cpp
