--- ViAn/table.h
@@ -1,11 +1,0 @@
-#ifndef TABLE_H
-#define TABLE_H
-
-
-class Table
-{
-public:
-    Table();
-};
-
-#endif // TABLE_H
